1. 支持CC字幕，可以显示视频的外挂字幕 2. 应用现已支持后台播放，你可以在设置中启用或关闭它 3. 视频打开时将获取云端历史，自动同步播放进度 4. 优化互动视频观看体验，左上角添加返回起点按钮 5. 允许设置主题跟随系统 6. 允许调整主副页排布方式断点（大于此值则主副页并排展示，小于则覆盖展示，你可以调整它以便在看视频的同时浏览评论） 7. 应用会记录你的清晰度调整记录 8. 新增关于弹幕控制条的设置项，允许自定义弹幕控制条在各种播放环境下的显隐 9. 修复部分动态加载失败的问题 10. 修复评论点赞UI没有正确反馈的问题

+++ BiliBili-Lib/Models/BiliBili/Topic.cs
@@ -35,13 +35,13 @@ namespace BiliBili_Lib.Models.BiliBili
         public int repost { get; set; }
         public int like { get; set; }
         public long dynamic_id { get; set; }
-        public int timestamp { get; set; }
-        public int pre_dy_id { get; set; }
-        public int orig_dy_id { get; set; }
+        public long timestamp { get; set; }
+        public long pre_dy_id { get; set; }
+        public long orig_dy_id { get; set; }
         public int uid_type { get; set; }
         public int stype { get; set; }
         public int r_type { get; set; }
-        public int inner_id { get; set; }
+        public long inner_id { get; set; }
         public int status { get; set; }
         public string dynamic_id_str { get; set; }
         public string pre_dy_id_str { get; set; }
